Some doc, refactoring to be able to launch the script account directly
