+++ tools/prod-scripts/jwt.js
