+++ tools/prod-scripts/mysql.js
